pyproject.toml
[project]
name = "mbti-reasoners"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llm-reasoners>=1.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
